requirements.txt
tensorflow
numpy
pandas
biopython
scikit-learn
simple-term-menu
